+++ packages/orq-rc/RELEASES.md
@@ -7959,3 +7959,13 @@ Based on:
 - [typescript v4.7.0-rc.39] packages/orq-rc
 ### Releases
 - [NPM v4.7.0-rc.39] https://www.npmjs.com/package/@orq-ai/node/v/4.7.0-rc.39 - packages/orq-rc
+
+## 2026-04-02 06:08:07
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.0 (2.879.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.7.0-rc.40] packages/orq-rc
+### Releases
+- [NPM v4.7.0-rc.40] https://www.npmjs.com/package/@orq-ai/node/v/4.7.0-rc.40 - packages/orq-rc